Refatoração das Classes do Transcritor Numérico.
