+++ src/Porter2023/Interfaces/ITranscritorNumericoInteiro.cs
@@ -1,0 +1,7 @@
+﻿namespace Porter2023.Interfaces
+{
+    public interface ITranscritorNumericoInteiro
+    {
+        string Transcrever(int numero);
+    }
+}
